+++ packages/conventional-changelog-conventionalcommits/src/utils.js
@@ -1,0 +1,11 @@
+export function matchScope(config = {}, commit) {
+  const {
+    scope: targetScope,
+    scopeOnly = false
+  } = config
+  const includesScope = commit.scope?.split(',').includes(targetScope)
+
+  return !targetScope
+    || (scopeOnly && includesScope)
+    || (!scopeOnly && (!commit.scope || includesScope))
+}