+++ packages/conventional-changelog/src/ConventionalChangelog.ts
@@ -21,7 +21,9 @@ import {
 import {
   type UnknownPresetCreatorParams,
   type PresetParams,
-  loadPreset
+  type PresetModuleLoader,
+  createPresetLoader,
+  loadPreset as defaultLoadPreset
 } from 'conventional-changelog-preset-loader'
 import normalizePackageData from 'normalize-package-data'
 import { findPackage } from 'fd-package-json'
@@ -386,11 +388,14 @@ export class ConventionalChangelog {
   /**
    * Load configs from a preset
    * @param preset
+   * @param loader - Preset module loader, if not provided, will use default loader
    * @returns this
    */
   loadPreset<PresetCreatorParams extends UnknownPresetCreatorParams = UnknownPresetCreatorParams>(
-    preset: PresetParams<PresetCreatorParams>
+    preset: PresetParams<PresetCreatorParams>,
+    loader?: PresetModuleLoader
   ) {
+    const loadPreset = loader ? createPresetLoader(loader) : defaultLoadPreset
     const config = loadPreset(preset).then((config) => {
       if (!config) {
         throw Error('Preset is not loaded or have incorrect exports')