+++ packages/git-client/src/GitClient.ts
@@ -1,15 +1,23 @@
+import { spawn } from 'child_process'
 import {
-  spawn,
-  stdoutSpawn,
-  splitStream,
-  getFirstFromStream,
+  firstFromStream,
+  splitStream
+} from '@simple-libs/stream-utils'
+import {
+  output,
+  outputStream
+} from '@simple-libs/child-process-utils'
+import {
   formatArgs,
   toArray
 } from './utils.js'
 import type {
   GitLogParams,
+  GitLogTagsParams,
   GitCommitParams,
   GitTagParams,
+  GitPushParams,
+  GitFetchParams,
   Arg
 } from './types.js'
 
@@ -35,6 +43,44 @@ export class GitClient {
   }
 
   /**
+   * Raw exec method to run git commands.
+   * @param args
+   * @returns Stdout string output of the command.
+   */
+  async exec(...args: Arg[]) {
+    return (
+      await output(spawn('git', this.formatArgs(...args), {
+        cwd: this.cwd
+      }))
+    ).toString().trim()
+  }
+
+  /**
+   * Raw exec method to run git commands with stream output.
+   * @param args
+   * @returns Stdout stream of the command.
+   */
+  execStream(...args: Arg[]) {
+    return outputStream(spawn('git', this.formatArgs(...args), {
+      cwd: this.cwd
+    }))
+  }
+
+  /**
+   * Initialize a new git repository.
+   * @returns Boolean result.
+   */
+  async init() {
+    try {
+      await this.exec('init')
+
+      return true
+    } catch (err) {
+      return false
+    }
+  }
+
+  /**
    * Get raw commits stream.
    * @param params
    * @param params.path - Read commits from specific path.
@@ -57,7 +103,7 @@ export class GitClient {
     const shouldNotIgnore = ignore
       ? (chunk: string) => !ignore.test(chunk)
       : () => true
-    const args = this.formatArgs(
+    const stdout = this.execStream(
       'log',
       `--format=${format}%n${SCISSOR}`,
       since && `--since=${since instanceof Date ? since.toISOString() : since}`,
@@ -67,9 +113,6 @@ export class GitClient {
       [from, to].filter(Boolean).join('..'),
       ...path ? ['--', ...toArray(path)] : []
     )
-    const stdout = stdoutSpawn('git', args, {
-      cwd: this.cwd
-    })
     const commitsStream = splitStream(stdout, `${SCISSOR}\n`)
     let chunk: string
 
@@ -82,19 +125,26 @@ export class GitClient {
 
   /**
    * Get tags stream.
+   * @param params
    * @yields Tags
    */
-  async* getTags() {
+  async* getTags(params: GitLogTagsParams = {}) {
+    const {
+      path,
+      from = '',
+      to = 'HEAD',
+      since
+    } = params
     const tagRegex = /tag:\s*(.+?)[,)]/gi
-    const args = this.formatArgs(
+    const stdout = this.execStream(
       'log',
       '--decorate',
       '--no-color',
-      '--date-order'
+      '--date-order',
+      since && `--since=${since instanceof Date ? since.toISOString() : since}`,
+      [from, to].filter(Boolean).join('..'),
+      ...path ? ['--', ...toArray(path)] : []
     )
-    const stdout = stdoutSpawn('git', args, {
-      cwd: this.cwd
-    })
     let chunk: Buffer
     let matches: IterableIterator<RegExpMatchArray>
     let tag: string
@@ -110,10 +160,11 @@ export class GitClient {
 
   /**
    * Get last tag.
+   * @param params
    * @returns Last tag, `null` if not found.
    */
-  async getLastTag() {
-    return getFirstFromStream(this.getTags())
+  async getLastTag(params?: GitLogTagsParams) {
+    return firstFromStream(this.getTags(params))
   }
 
   /**
@@ -122,16 +173,12 @@ export class GitClient {
    * @returns Boolean value.
    */
   async checkIgnore(file: string) {
-    const args = this.formatArgs(
-      'check-ignore',
-      '--',
-      file
-    )
-
     try {
-      await spawn('git', args, {
-        cwd: this.cwd
-      })
+      await this.exec(
+        'check-ignore',
+        '--',
+        file
+      )
 
       return true
     } catch (err) {
@@ -144,15 +191,11 @@ export class GitClient {
    * @param files - Files to stage.
    */
   async add(files: string | string[]) {
-    const args = this.formatArgs(
+    await this.exec(
       'add',
       '--',
       ...toArray(files)
     )
-
-    await spawn('git', args, {
-      cwd: this.cwd
-    })
   }
 
   /**
@@ -161,6 +204,7 @@ export class GitClient {
    * @param params.verify
    * @param params.sign
    * @param params.files
+   * @param params.allowEmpty
    * @param params.message
    */
   async commit(params: GitCommitParams) {
@@ -168,21 +212,20 @@ export class GitClient {
       verify = true,
       sign = false,
       files = [],
+      allowEmpty = false,
       message
     } = params
-    const args = this.formatArgs(
+
+    await this.exec(
       'commit',
       !verify && '--no-verify',
       sign && '-S',
+      allowEmpty && '--allow-empty',
       '-m',
       message,
       '--',
       ...files
     )
-
-    await spawn('git', args, {
-      cwd: this.cwd
-    })
   }
 
   /**
@@ -203,7 +246,7 @@ export class GitClient {
       message = ''
     }
 
-    const args = this.formatArgs(
+    await this.exec(
       'tag',
       sign && '-s',
       message && '-a',
@@ -211,10 +254,6 @@ export class GitClient {
       '--',
       name
     )
-
-    await spawn('git', args, {
-      cwd: this.cwd
-    })
   }
 
   /**
@@ -222,16 +261,27 @@ export class GitClient {
    * @returns Current branch name.
    */
   async getCurrentBranch() {
-    const args = this.formatArgs(
+    const branch = await this.exec(
       'rev-parse',
       '--abbrev-ref',
       'HEAD'
     )
+
+    return branch
+  }
+
+  /**
+   * Get default branch name.
+   * @returns Default branch name.
+   */
+  async getDefaultBranch() {
     const branch = (
-      await spawn('git', args, {
-        cwd: this.cwd
-      })
-    ).toString().trim()
+      await this.exec(
+        'rev-parse',
+        '--abbrev-ref',
+        'origin/HEAD'
+      )
+    ).replace(/^origin\//, '')
 
     return branch
   }
@@ -239,38 +289,50 @@ export class GitClient {
   /**
    * Push changes to remote.
    * @param branch
+   * @param params
+   * @param params.verify
    */
-  async push(branch: string) {
-    const args = this.formatArgs(
+  async push(
+    branch: string,
+    params: GitPushParams = {}
+  ) {
+    const {
+      verify = true,
+      tags = false,
+      followTags = false,
+      force = false
+    } = params
+
+    await this.exec(
       'push',
-      '--follow-tags',
+      followTags && '--follow-tags',
+      tags && '--tags',
+      !verify && '--no-verify',
+      force && '--force',
       'origin',
       '--',
       branch
     )
-
-    await spawn('git', args, {
-      cwd: this.cwd
-    })
   }
 
   /**
    * Verify rev exists.
    * @param rev
+   * @param safe - If `true`, will not throw error if rev not found.
    * @returns Target hash.
    */
-  async verify(rev: string) {
-    const args = this.formatArgs(
+  async verify(rev: string, safe?: boolean) {
+    let git = this.exec(
       'rev-parse',
       '--verify',
       rev
     )
 
-    return (
-      await spawn('git', args, {
-        cwd: this.cwd
-      })
-    ).toString().trim()
+    if (safe) {
+      git = git.catch(() => '')
+    }
+
+    return await git
   }
 
   /**
@@ -279,17 +341,89 @@ export class GitClient {
    * @returns Config value.
    */
   async getConfig(key: string) {
-    const args = this.formatArgs(
+    return await this.exec(
       'config',
       '--get',
       '--',
       key
     )
+  }
 
-    return (
-      await spawn('git', args, {
-        cwd: this.cwd
-      })
-    ).toString().trim()
+  /**
+   * Set config value by key.
+   * @param key - Config key.
+   * @param value - Config value.
+   */
+  async setConfig(key: string, value: string) {
+    await this.exec(
+      'config',
+      '--',
+      key,
+      value
+    )
+  }
+
+  /**
+   * Fetch changes from remote.
+   * @param params
+   */
+  async fetch(params: GitFetchParams = {}) {
+    const {
+      prune = false,
+      unshallow = false,
+      tags = false,
+      all = false,
+      remote,
+      branch
+    } = params
+
+    await this.exec(
+      'fetch',
+      prune && '--prune',
+      unshallow && '--unshallow',
+      tags && '--tags',
+      all && '--all',
+      ...remote && branch ? [
+        '--',
+        remote,
+        branch
+      ] : []
+    )
+  }
+
+  /**
+   * Create a new branch.
+   * @param branch - Branch name.
+   */
+  async createBranch(branch: string) {
+    await this.exec(
+      'checkout',
+      '-b',
+      branch
+    )
+  }
+
+  /**
+   * Delete a branch.
+   * @param branch - Branch name.
+   */
+  async deleteBranch(branch: string) {
+    await this.exec(
+      'branch',
+      '-D',
+      '--',
+      branch
+    )
+  }
+
+  /**
+   * Checkout a branch.
+   * @param branch - Branch name.
+   */
+  async checkout(branch: string) {
+    await this.exec(
+      'checkout',
+      branch
+    )
   }
 }