+++ packages/git-client/src/index.ts
@@ -1,3 +1,4 @@
 export * from './types.js'
+export * from './utils.js'
 export * from './GitClient.js'
 export * from './ConventionalGitClient.js'